# Speed up adaptive threshold in findChessboardCorners