--- a/modules/calib3d/src/calibinit.cpp
+++ b/modules/calib3d/src/calibinit.cpp
@@ -583,10 +583,10 @@ bool findChessboardCorners(InputArray image_, Size pattern_size,
         }
         //if flag CALIB_CB_ADAPTIVE_THRESH is not set it doesn't make sense to iterate over k
         int max_k = useAdaptive ? 6 : 1;
+        Mat prev_thresh_img;
         for (int k = 0; k < max_k && !found; k++)
         {
             int prev_block_size = -1;
-            Mat prev_thresh_img;
             for (int dilations = min_dilations; dilations <= max_dilations; dilations++)
             {
                 // convert the input grayscale image to binary (black-n-white)
@@ -603,10 +603,9 @@ bool findChessboardCorners(InputArray image_, Size pattern_size,
                         dilate( thresh_img, thresh_img, Mat(), Point(-1, -1), dilations );
                         thresh_img.copyTo(prev_thresh_img);
                     }
-                    else
+                    else if (dilations > 0)
                     {
-                        if (dilations > 0)
-                            dilate( prev_thresh_img, prev_thresh_img, Mat(), Point(-1, -1), 1 );
+                        dilate( prev_thresh_img, prev_thresh_img, Mat(), Point(-1, -1), 1 );
                         prev_thresh_img.copyTo(thresh_img);
                     }
                     prev_block_size = block_size;