@AleksandrPanov 2024-03-27
> @AleksandrPanov Moving `int prev_block_size = -1;` outside from loop on variable `k` should not be equivalent change. Variable `k` is used as argument in `adaptiveThreshold` function, so we shouldn't reuse result from different `k` - only for same `k`. This change decreased detected chessboards number
> 
> ```
> cell_img_size = 100 (default)
> 
>                                  category  detected chessboard  total detected chessboard  total chessboard  average detected error chessboard
>                                       all             0.469306                       6758             14400                           4.178116
> 
> ----------------------------------------------------------------------------------------------------------------------------------------------
> 
> cell_img_size = 10
> 
>                                  category  detected chessboard  total detected chessboard  total chessboard  average detected error chessboard
>                                       all             0.895833                      12900             14400                           0.600768
> ```

@MaximSmolskiy, thx, fixed