@AleksandrPanov 2024-03-27
Now the results are correct:
`--configuration=generate_run --board_x=7 --path=res_chessboard --synthetic_object=chessboard`
![image](https://github.com/opencv/opencv/assets/22337800/b4daa8f4-58d4-42d4-8753-8501f08cedd5)

`--configuration=generate_run --board_x=7 --path=res_chessboard --synthetic_object=chessboard --cell_img_size=10`
![image](https://github.com/opencv/opencv/assets/22337800/2af195d6-11dc-4f0e-a6ad-411e8fa3256d)